Change BanStick to be built automatically instead of manually
This is an example of how the rest of the plugins may be
automatically built from source. This approach means plugins do not
need to be manually compiled, and the plugin jars do not need to be
checked in to version control.

Updating plugins can also be automated by bumping the version and
SHA256 hash to the newer version. It is important to check the
SHA256 hash in order to verify the integrity of the source code
matches the tag we are trying to build.

Unforutunately many of the plugins don't seem to provide binaries,
so we must build the plugin jar ourselves. This requires Maven
packages from the GitHub Package Repository which requires a valid
Personal Access Token with read:packages.

It would be even nicer if the upstream plugins could provide their
own compiled jars, and all CivDocker had to do was download the jar
from the plugin instead of downloading the plugin's source code and
building. This would also avoid having to provide a GitHub Personal
Access Token.

It would also be nice to make sure these builds are reproducible,
but that is something to think about later.
See https://reproducible-builds.org/ and
https://reproducible-builds.org/docs/jvm/